move utils to the module
